WIP genomic region trait issue
